# [runtime] optimize `strtolower` and `strtoupper`